ヨシ！

+++ src/main/java/org/contan_lang/environment/CoroutineStatus.java
@@ -4,11 +4,11 @@ import org.contan_lang.variables.ContanObject;
 
 public class CoroutineStatus {
 
-    public final int count;
+    public final long count;
     
     public final ContanObject<?>[] results;
     
-    public CoroutineStatus(int count, ContanObject<?>... results) {
+    public CoroutineStatus(long count, ContanObject<?>... results) {
         this.count = count;
         this.results = results;
     }
